requirements.txt
aiofiles==23.2.1
aiogram==3.4.1
aiohttp==3.9.3
aiohttp-socks==0.8.4
aiosignal==1.3.1
aiosqlite==0.20.0
annotated-types==0.6.0
anyio==4.3.0
async-timeout==4.0.3
attrs==23.2.0
certifi==2024.2.2
cssselect==1.2.0
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.6
jmespath==1.0.1
lxml==5.2.1
magic-filter==1.0.12
multidict==6.0.5
packaging==24.0
parsel==1.9.1
pydantic==2.5.3
pydantic_core==2.14.6
python-dotenv==1.0.1
python-socks==2.4.4
sniffio==1.3.1
typing_extensions==4.10.0
w3lib==2.1.2
yarl==1.9.4
